+++ src/main/java/com/github/nginate/kafka/protocol/messages/Message.java
@@ -3,10 +3,10 @@ package com.github.nginate.kafka.protocol.messages;
 import com.github.nginate.kafka.protocol.types.Type;
 import lombok.Data;
 
-import static com.github.nginate.kafka.protocol.types.TypeName.INT8;
+import static com.github.nginate.kafka.protocol.types.TypeName.INT32;
 
 @Data
 public abstract class Message {
-    @Type(INT8)
-    private Byte size;
+    @Type(INT32)
+    private Integer size;
 }
